requirements.txt
groq>=0.9.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
gradio>=4.0.0
scikit-learn>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
numpy>=1.24.0
